+++ src/id_translation/exceptions.py
@@ -1,4 +1,5 @@
 """General errors for the translation suite."""
+from warnings import simplefilter as _simplefilter
 
 
 class ConfigurationError(ValueError):
@@ -15,3 +16,14 @@ class TranslationError(ValueError):
 
 class TooManyFailedTranslationsError(TranslationError):
     """Raised if too many IDs fail to translate."""
+
+
+class TranslationWarning(UserWarning):
+    """Base class for translation warnings."""
+
+
+class TranslationDisabledWarning(TranslationWarning):
+    """Translation is globally disabled."""
+
+
+_simplefilter("once", category=TranslationDisabledWarning)